+++ tests/test_checks_googlefonts.py
@@ -2562,38 +2562,6 @@ def test_condition_familyname_with_spaces():
     )
 
 
-# TODO: Maybe skip this code-test if the service is offline?
-# we could use pytest.mak.skipif here together with a piece of code that
-# verifies whether or not the namecheck.fontdata.com website is online at the moment
-def test_check_fontdata_namecheck():
-    """Familyname is unique according to namecheck.fontdata.com"""
-    check = CheckTester("fontdata_namecheck")
-
-    TIMEOUT_MSG = (
-        "Sometimes namecheck.fontdata.com times out"
-        " and we don't want to stop running all the other"
-        " code tests. Unless you touched this portion of"
-        " the code, it is generaly safe to ignore this glitch."
-    )
-    # We dont FAIL because this is meant as a merely informative check
-    # There may be frequent cases when fonts are being updated and thus
-    # already have a public family name registered on the
-    # namecheck.fontdata.com database.
-    font = TEST_FILE("cabin/Cabin-Regular.ttf")
-    assert_results_contain(
-        check(font),
-        INFO,
-        "name-collision",
-        "with an already used name...",
-        ignore_error=TIMEOUT_MSG,
-    )
-
-    # Here we know that FamilySans has not been (and will not be)
-    # registered as a real family.
-    font = TEST_FILE("familysans/FamilySans-Regular.ttf")
-    assert_PASS(check(font), "with a unique family name...", ignore_error=TIMEOUT_MSG)
-
-
 # Temporarily disabling this code-test since check/negative_advance_width itself
 # is disabled waiting for an implementation targetting the
 # actual root cause of the issue.
